+++ LanguageCenterPLC.Data/Entities/Timesheet.cs
@@ -83,7 +83,7 @@ namespace LanguageCenterPLC.Data.Entities
         public float Day_31 { get; set; }
         #endregion
 
-        public int TotalWorkday { get; set; }
+        public float TotalWorkday { get; set; }
 
         public decimal Salary { get; set; }
